+++ scripts/build_camino.sh
@@ -1,15 +1,29 @@
 #!/usr/bin/env bash
 
-set -o errexit
-set -o nounset
-set -o pipefail
+set -euo pipefail
 
-echo "Building camino node..."
+print_usage() {
+  printf "Usage: build_camino [OPTIONS]
+
+  Build caminogo
+
+  Options:
+
+    -r  Build with race detector
+"
+}
+
+race=''
+while getopts 'r' flag; do
+  case "${flag}" in
+    r) race='-race' ;;
+    *) print_usage
+      exit 1 ;;
+  esac
+done
 
 # Changes to the minimum golang version must also be replicated in
-# scripts/ansible/roles/golang_base/defaults/main.yml (here)
 # scripts/build_camino.sh (here)
-# scripts/local.Dockerfile
 # Dockerfile
 # README.md
 # go.mod
@@ -32,7 +46,7 @@ version_lt() {
 }
 
 if version_lt "$(go_version)" "$go_version_minimum"; then
-    echo "Camino-Node requires Go >= $go_version_minimum, Go $(go_version) found." >&2
+    echo "CaminoGo requires Go >= $go_version_minimum, Go $(go_version) found." >&2
     exit 1
 fi
 
@@ -48,14 +62,16 @@ LDFLAGS="$LDFLAGS -X github.com/ava-labs/coreth/plugin/evm.GitCommit=$caminoethv
 LDFLAGS="$LDFLAGS -X github.com/ava-labs/coreth/plugin/evm.Version=$caminoethvm_tag"
 LDFLAGS="$LDFLAGS $static_ld_flags"
 
-go build -ldflags "$LDFLAGS" -o "$caminogo_path" "$CAMINOGO_PATH/main/"*.go
+build_args="$race"
+echo "Building CaminoGo..."
+go build $build_args -ldflags "$LDFLAGS" -o "$CAMINOGO_BIN_PATH" "$CAMINOGO_PATH/main/"*.go
 
 # Make plugin folder
-mkdir -p $plugin_dir
+mkdir -p "$plugin_dir"
 
 # Exit build successfully if the binaries are created
-if [[ -f "$caminogo_path" ]]; then
-    ln -sf caminogo $camino_node_symlink_path
+if [[ -f "$CAMINOGO_BIN_PATH" ]]; then
+    ln -sf caminogo "$camino_node_symlink_path"
     echo "Build Successful"
     exit 0
 else