+++ scripts/build_publish_image.sh
@@ -8,7 +8,7 @@ set -o pipefail
 CAMINOGO_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )"; cd .. && pwd )
 
 # Build the container
-"$CAMINOGO_PATH"/scripts/build_local_image.sh
+"$CAMINOGO_PATH"/scripts/build_image.sh
 
 # If this is not a trusted build (Docker Credentials are not set)
 if [[ -z "$DOCKER_USERNAME"  ]]; then
@@ -23,4 +23,4 @@ echo "Pushing: $camino_node_dockerhub_repo:$current_branch"
 echo "$DOCKER_PASS" | docker login --username "$DOCKER_USERNAME" --password-stdin
 
 ## pushing image with tags
-docker image push -a $camino_node_dockerhub_repo
+docker image push -a "$camino_node_dockerhub_repo"