+++ scripts/constants.sh
@@ -1,15 +1,18 @@
 #!/usr/bin/env bash
-#
+
+# Ignore warnings about variables appearing unused since this file is not the consumer of the variables it defines.
+# shellcheck disable=SC2034
+
+set -euo pipefail
+
 # Use lower_case variables in the scripts and UPPER_CASE variables for override
 # Use the constants.sh for env overrides
 
-# Set the PATHS
-GOPATH="$(go env GOPATH)"
 CAMINOGO_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )"; cd .. && pwd )
 
 # Where CaminoGo binary goes
 build_dir="$CAMINOGO_PATH/build"
-caminogo_path="$build_dir/caminogo"
+CAMINOGO_BIN_PATH="$build_dir/caminogo"
 camino_node_symlink_path="$build_dir/camino-node"
 plugin_dir="$build_dir/plugins"