+++ scripts/test.sh
@@ -13,4 +13,4 @@ export CGO_CFLAGS="-O -D__BLST_PORTABLE__"
 # clear error due to the default value change in go1.20.
 export CGO_ENABLED=1
 
-go test ${1-} -timeout="120s" -coverprofile="coverage.out" -covermode="atomic" $(go list ./... | grep -v /mocks | grep -v proto | grep -v tests)
+go test "${1-}" -timeout="120s" -coverprofile="coverage.out" -covermode="atomic" "$(go list ./... | grep -v /mocks | grep -v proto | grep -v tests)"