+++ scripts/tests.e2e.sh
@@ -1,11 +1,10 @@
 #!/usr/bin/env bash
-set -e
-set -o nounset
-set -o pipefail
+
+set -euo pipefail
 
 # e.g.,
 # ./scripts/build.sh
-# ./scripts/tests.e2e.sh ./build/caminogo
+# ./scripts/tests.e2e.sh ./build/caminogo ./tools/camino-network-runner/bin/camino-network-runner
 # ENABLE_WHITELIST_VTX_TESTS=false ./scripts/tests.e2e.sh ./build/caminogo ./tools/camino-network-runner/bin/camino-network-runner
 if ! [[ "$0" =~ scripts/tests.e2e.sh ]]; then
   echo "must be run from repository root"
@@ -14,17 +13,17 @@ fi
 
 echo "Run caminogo e2e tests..."
 
-CAMINOGO_PATH="${1-}"
-if [[ -z "${CAMINOGO_PATH}" ]]; then
+CAMINOGO_BIN_PATH="${1-}"
+if [[ -z "${CAMINOGO_BIN_PATH}" ]]; then
   echo "Missing CAMINOGO_PATH argument!"
-  echo "Usage: ${0} [CAMINOGO_PATH]" >> /dev/stderr
+  echo "Usage: ${0} [CAMINOGO_BIN_PATH] [CAMINO_NETWORK_RUNNER_PATH]" >> /dev/stderr
   exit 255
 fi
 
 CAMINO_NETWORK_RUNNER_PATH="${2-}"
 if [[ -z "${CAMINO_NETWORK_RUNNER_PATH}" ]]; then
   echo "Missing CAMINO_NETWORK_RUNNER_PATH argument!"
-  echo "Usage: ${0} [CAMINOGO_PATH] [CAMINO_NETWORK_RUNNER_PATH]" >> /dev/stderr
+  echo "Usage: ${0} [CAMINOGO_BIN_PATH] [CAMINO_NETWORK_RUNNER_PATH]" >> /dev/stderr
   exit 255
 fi
 
@@ -37,7 +36,7 @@ if [[ ${ENABLE_WHITELIST_VTX_TESTS} == true ]]; then
 fi
 echo GINKGO_LABEL_FILTER: ${GINKGO_LABEL_FILTER}
 
-cp $CAMINO_NETWORK_RUNNER_PATH /tmp/camino-network-runner
+cp "$CAMINO_NETWORK_RUNNER_PATH" /tmp/camino-network-runner
 
 GOPATH="$(go env GOPATH)"
 PATH="${GOPATH}/bin:${PATH}"
@@ -74,12 +73,12 @@ PID=${!}
 # --ginkgo.skip "\[Local\]"
 #
 # set "--enable-whitelist-vtx-tests" to explicitly enable/disable whitelist vtx tests
-echo "running e2e tests against the local cluster with ${CAMINOGO_PATH}"
+echo "running e2e tests against the local cluster with ${CAMINOGO_BIN_PATH}"
 ./tests/e2e/e2e.test \
 --ginkgo.v \
 --log-level debug \
 --network-runner-grpc-endpoint="0.0.0.0:12342" \
---network-runner-camino-node-path=${CAMINOGO_PATH} \
+--network-runner-camino-node-path="${CAMINOGO_BIN_PATH}" \
 --network-runner-camino-log-level="WARN" \
 --test-keys-file=tests/test.insecure.secp256k1.keys --ginkgo.label-filter="${GINKGO_LABEL_FILTER}" \
 && EXIT_CODE=$? || EXIT_CODE=$?
@@ -88,7 +87,7 @@ kill ${PID}
 
 if [[ ${EXIT_CODE} -gt 0 ]]; then
   echo "FAILURE with exit code ${EXIT_CODE}"
-  exit ${EXIT_CODE}
+  exit "${EXIT_CODE}"
 else
   echo "ALL SUCCESS!"
 fi